extract applyEnvVars into separate module
